+++ test/components/avatar_component_test.rb
@@ -26,14 +26,20 @@ class PrimerAvatarComponentTest < Minitest::Test
   def test_defaults_adds_small_modifier_when_size_is_less_than_threshold
     render_inline(Primer::AvatarComponent.new(src: "https://github.com/github.png", alt: "github", size: Primer::AvatarComponent::SMALL_THRESHOLD - 1))
 
-    assert_selector("img.avatar.avatar--small")
+    assert_selector("img.avatar.avatar-small")
   end
 
   def test_defaults_does_not_add_small_modifier_when_size_is_greater_than_threshold
     render_inline(Primer::AvatarComponent.new(src: "https://github.com/github.png", alt: "github", size: Primer::AvatarComponent::SMALL_THRESHOLD + 1))
 
     assert_selector("img.avatar")
-    refute_selector(".avatar--small")
+    refute_selector(".avatar-small")
+  end
+
+  def test_adds_custom_classes
+    render_inline(Primer::AvatarComponent.new(src: "https://github.com/github.png", alt: "github", classes: "custom-class"))
+
+    assert_selector("img.avatar.custom-class")
   end
 
   def test_sets_size_height_and_width
@@ -46,16 +52,71 @@ class PrimerAvatarComponentTest < Minitest::Test
     render_inline(Primer::AvatarComponent.new(src: "https://github.com/github.png", alt: "github", square: true))
 
     assert_selector("img.avatar")
-    refute_selector(".CircleBadge")
+    refute_selector(".circle")
   end
 
   def test_renders_link_wrapper
+    render_inline(Primer::AvatarComponent.new(src: "https://github.com/github.png", alt: "github", href: "#given-href"))
+
+    assert_selector("a.avatar") do |(a)|
+      assert_equal("#given-href", a["href"])
+      assert_selector("img")
+      refute_selector("img.avatar")
+    end
+  end
+
+  def test_defaults_circle_link_wrapper
     render_inline(Primer::AvatarComponent.new(src: "https://github.com/github.png", alt: "github", href: "#"))
+
+    assert_selector("a.avatar.circle") do
+      assert_selector("img")
+      refute_selector("img.circle")
+    end
+  end
+
+  def test_squared_link_wrapper
+    render_inline(Primer::AvatarComponent.new(src: "https://github.com/github.png", alt: "github", href: "#", square: true))
 
     assert_selector("a.avatar") do
       assert_selector("img")
     end
-    refute_selector("img.avatar")
+    refute_selector(".circle")
+  end
+
+  def test_adds_small_modifier_to_link_wrapper_when_size_is_less_than_threshold
+    render_inline(Primer::AvatarComponent.new(src: "https://github.com/github.png", alt: "github", href: "#", size: Primer::AvatarComponent::SMALL_THRESHOLD - 1))
+
+    assert_selector("a.avatar.avatar-small") do
+      assert_selector("img")
+      refute_selector("img.avatar-small")
+    end
+  end
+
+  def test_does_not_add_small_modifier_to_link_wrapper_when_size_is_greater_than_threshold
+    render_inline(Primer::AvatarComponent.new(src: "https://github.com/github.png", alt: "github", href: "#", size: Primer::AvatarComponent::SMALL_THRESHOLD + 1))
+
+    assert_selector("a.avatar") do
+      assert_selector("img")
+    end
+    refute_selector(".avatar-small")
+  end
+
+  def test_adds_custom_classes_to_link_wrapper
+    render_inline(Primer::AvatarComponent.new(src: "https://github.com/github.png", alt: "github", href: "#", classes: "custom-class"))
+
+    assert_selector("a.avatar.custom-class") do
+      assert_selector("img")
+      refute_selector("img.custom-class")
+    end
+  end
+
+  def test_clears_link_wrapper_line_height
+    render_inline(Primer::AvatarComponent.new(src: "https://github.com/github.png", alt: "github", href: "#"))
+
+    assert_selector("a.lh-0") do
+      assert_selector("img")
+      refute_selector("img.lh-0")
+    end
   end
 
   def test_status